Godrej Indiranagar Extension Bangalore
Godrej Indiranagar Extension Bangalore is latest launch residential project developed by well-known developer Godrej Properties. As per best floor plan 2 BHK, 3 BHK, and 4 BHK luxury residential accommodation.

The project is destined to cater to a life full of luxury in Indiranagar Bangalore, which is the fastest-developing destination in the city. It is going to be a perfect blend of grandeur and fresh living with best-in-class features and tremendous greenery around. Also, the proximity to CMH Road and the proposed Purple Line of the Namma Metro adds more charm to its aspects and the Kempegowda International Airport is also at a distance of just a few minutes from this location.

Godrej Indiranagar provides best living experience for you and your family in Bengaluru. These include a themed kids’ play area, an seniors’ corner, a walking & jogging path, an outdoor gym, a pavilion, a function lawn, urban farming, a pet’s corner with play mounds, and a swimming pool & kids’ pool.

Facing panoramic views and having exceptional natural beauty in the form of flora and fauna, ensures a healthy and fruitful life for the residents. It is considered a source of happiness for people of all ages such as the field of children playing for children, for children to play their hearts happily, a meditation center to relax the souls of big people, stay fit For fully equipped gymnasiums and swimming pools. 

Godrej Indiranagar Extension is located in a prime location close to the intersection of the Chinmaya Mission Hospital Rd which can be easily accessible to the major areas of Bangalore as well as famous malls, schools, and hospitals.

For More Details
Visit us - https://bit.ly/3DtisKc